+++ packages/starlight-view-modes/virtual.d.ts
@@ -1,11 +1,11 @@
-declare module "virtual:starlight-view-modes-config" {
-  const StarlightViewModesConfig: import("./index").StarlightViewModesConfig;
+declare module "virtual:starlight-view-modes/config" {
+  const StarlightViewModesConfig: import("./libs/config").StarlightViewModesConfig;
+
   export default StarlightViewModesConfig;
 }
 
-declare module "virtual:starlight-view-modes-context" {
+declare module "virtual:starlight-view-modes/context" {
   const StarlightViewModesContext: import("./libs/vite").StarlightViewModesContext;
+
   export default StarlightViewModesContext;
 }
-
-declare module "picomatch";